refactor: tag로 이름 변경 (#447)
